requirements.txt
attrs==22.1.0
aws-cdk-lib==2.51.1
aws-cdk.asset-awscli-v1==2.2.16
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.22
aws-cdk.aws-sagemaker-alpha==2.51.1a0
bcrypt==4.0.1
boto3==1.26.15
botocore==1.29.15
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constructs==10.1.168
cryptography==38.0.4
exceptiongroup==1.0.4
gevent==22.10.2
graphviz==0.8.4
greenlet==2.0.1
idna==3.4
inotify-simple==1.2.1
jmespath==1.0.1
joblib==1.2.0
jsii==1.71.0
MarkupSafe==2.1.1
mxnet==1.9.1
numpy==1.23.5
pandas==1.5.2
paramiko==2.12.0
protobuf==3.19.6
psutil==5.9.4
publication==0.0.3
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
retrying==1.3.3
s3transfer==0.6.0
sagemaker-mxnet-training==4.3.0
sagemaker-training==4.4.0
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.2
